Create QC - Find HC Asset Errors
--BETA VERSION--- limited testing only
This script is designed to find cases on which there have been errors with Health Care assets.  It finds all active HC cases with an asset test and an approval in the previous month, checks all asset panels, and adds those cases with counted assets that are not verified, or the verification was more than 30 days prior to the sending of the HC renewal, and adds to an excel spreadsheet.  
It also adds cases with ELIG type MA/DX and method X approved after 5/31/17 to the list.
Future enhancement planned to seek cases that were renewals or applications only.